Moved add_bool_arg from train.py to parsing_utils
